requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
azure-identity==1.23.0
azure-cosmos==4.9.0
azure-storage-blob==12.25.1
